Use remote fixtures repo
* restructure fixtures
* adapt code
* clean up unused code
* use application settings in the test code
* get rid of @current tag in the tests setup
