+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "saleor"]
 	path = saleor
 	url = git@github.com:mirumee/saleor
+[submodule "features/fixtures"]
+	path = features/fixtures
+	url = git@github.com:softozor/shopozor-fixtures
