--- features/fixtures/graphql/login.graphql
@@ -1,15 +1,0 @@
-mutation Login($email: String!, $password: String!, $isStaff: Boolean = false) {
-  login(email: $email, password: $password, isStaff: $isStaff) {
-    token
-    errors {
-      field
-      message
-    }
-    user {
-      permissions {
-        code
-      }
-      isStaff
-    }
-  }
-}
