+++ setup.cfg
@@ -40,4 +40,3 @@ split_before_first_argument = true
 paths = features
 verbose = true
 color = true
-tags = @current
