+++ shopozor/urls.py
@@ -1,6 +1,4 @@
 from django.conf import settings
-from django.conf.urls.static import static
-from django.contrib.staticfiles.views import serve
 from django.conf.urls import include, url
 from django.views.decorators.csrf import csrf_exempt
